Added macro for parameterized templates inclusion
Add the `<@include .../>` macro to `standard.ftl` that wraps the
built-in `<#include ...>` directive and allows to include parameterized
templates.

Fixed the default PLSQL template loader. Now it returns `null` when the
sought template is not found, and FreeMarker raises a proper exception
itself. Otherwise the `ignore_missing` option in the `<#include ...>`
directive doesn't work.